Moving data objects into the commercial package
