+++ commercial/app/controllers/commercial/BookOffersController.scala
@@ -1,8 +1,8 @@
 package controllers.commercial
 
 import common.{JsonComponent, ExecutionContexts, Logging}
-import model.commercial.Segment
-import model.commercial.books.{BestsellersAgent, Book, BookFinder, CacheNotConfiguredException}
+import model.commercial.{Segment, Book}
+import model.commercial.books.{BestsellersAgent, BookFinder, CacheNotConfiguredException}
 import model.commercial.{FeedMissingConfigurationException, FeedSwitchOffException}
 import model.{Cached, NoCache}
 import play.api.mvc._
