+++ commercial/app/controllers/commercial/SoulmatesController.scala
@@ -1,8 +1,10 @@
 package controllers.commercial
 
 import common.JsonComponent
+import model.commercial.soulmates.{SoulmatesAgent, Sample}
 import model.commercial.soulmates.SoulmatesAgent.{menAgent, newMenAgent, newWomenAgent, womenAgent}
 import model.commercial.soulmates._
+import model.commercial.Member
 import model.Cached
 import play.api.mvc._
