+++ commercial/app/model/commercial/soulmates/SoulmatesAgent.scala
@@ -2,6 +2,7 @@ package model.commercial.soulmates
 
 import commercial.feeds.{FeedMetaData, ParsedFeed}
 import common.AkkaAgent
+import model.commercial.{Member, Man, Woman}
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.Random
