+++ commercial/app/model/commercial/travel/TravelOffersAgent.scala
@@ -3,7 +3,7 @@ package model.commercial.travel
 import commercial.feeds.{FeedMetaData, ParsedFeed}
 import common.ExecutionContexts
 import contentapi.ContentApiClient
-import model.commercial.{Keyword, Lookup, MerchandiseAgent, Segment}
+import model.commercial.{Keyword, Lookup, MerchandiseAgent, Segment, TravelOffer}
 
 import scala.concurrent.Future
