+++ commercial/app/views/books/book.scala.html
@@ -1,4 +1,4 @@
-@(book: model.commercial.books.Book, omnitureId: Option[String], clickMacro: Option[String])(implicit request: RequestHeader)
+@(book: model.commercial.Book, omnitureId: Option[String], clickMacro: Option[String])(implicit request: RequestHeader)
 @import views.support.{ImgSrc, Item300}
 
 <aside class="adverts adverts--manual adverts--legacy-inline adverts--tone-books advert--book" data-link-name="merchandising | book | @omnitureId">
