+++ commercial/app/views/books/bookCard.scala.html
@@ -1,4 +1,4 @@
-@(book: model.commercial.books.Book,
+@(book: model.commercial.Book,
   clickMacro: Option[String],
   optClassNames: Option[Seq[String]] = None)
