+++ commercial/app/views/books/bookLargeCard.scala.html
@@ -1,4 +1,4 @@
-@(book: model.commercial.books.Book,
+@(book: model.commercial.Book,
   clickMacro: Option[String],
   optAdvertClassNames: Option[Seq[String]] = None)
