+++ commercial/app/views/books/booksBlended.scala.html
@@ -1,4 +1,4 @@
-@(book: model.commercial.books.Book, clickMacro: Option[String])
+@(book: model.commercial.Book, clickMacro: Option[String])
 
 <div class="advert-blended advert-blended--books">
     <div class="advert-blended__title">
