+++ commercial/app/views/books/booksStandard.scala.html
@@ -1,4 +1,4 @@
-@(books: Seq[model.commercial.books.Book],
+@(books: Seq[model.commercial.Book],
   omnitureId: Option[String],
   clickMacro: Option[String],
   isProminent: Boolean = false)(implicit request: RequestHeader)
