+++ commercial/app/views/jobs/jobCard.scala.html
@@ -1,4 +1,4 @@
-@(job: model.commercial.jobs.Job,
+@(job: model.commercial.Job,
   clickMacro: Option[String],
   optClassNames: Option[Seq[String]] = None)
