+++ commercial/app/views/jobs/jobLargeCard.scala.html
@@ -1,4 +1,4 @@
-@(job: model.commercial.jobs.Job,
+@(job: model.commercial.Job,
   clickMacro: Option[String],
   optAdvertClassNames: Option[Seq[String]] = None)
