+++ commercial/app/views/jobs/jobsBlended.scala.html
@@ -1,4 +1,4 @@
-@(job: model.commercial.jobs.Job, clickMacro: Option[String])
+@(job: model.commercial.Job, clickMacro: Option[String])
 
 <div class="advert-blended advert-blended--jobs">
     <div class="advert-blended__title">
