+++ commercial/app/views/liveevents/liveEvent.scala.html
@@ -1,4 +1,4 @@
-@(event: model.commercial.events.LiveEvent, omnitureId: Option[String], clickMacro: Option[String])(implicit request: RequestHeader)
+@(event: model.commercial.LiveEvent, omnitureId: Option[String], clickMacro: Option[String])(implicit request: RequestHeader)
 @import views.support.{ImgSrc, Item300}
 
     <aside class="adverts adverts--manual adverts--legacy-inline adverts--tone-membership adverts--membership" data-link-name="merchandising | event | @omnitureId">
