+++ commercial/app/views/masterclasses/masterclassCard.scala.html
@@ -1,4 +1,4 @@
-@(event: model.commercial.events.Masterclass,
+@(event: model.commercial.Masterclass,
   clickMacro: Option[String],
   optClassNames: Option[Seq[String]] = None)
