+++ commercial/app/views/masterclasses/masterclassLargeCard.scala.html
@@ -1,4 +1,4 @@
-@(event: model.commercial.events.Masterclass,
+@(event: model.commercial.Masterclass,
   clickMacro: Option[String],
   optAdvertClassNames: Option[Seq[String]] = None)
