+++ commercial/app/views/masterclasses/masterclasses.scala.html
@@ -1,4 +1,4 @@
-@(events: Seq[model.commercial.events.Masterclass],
+@(events: Seq[model.commercial.Masterclass],
   omnitureId: Option[String],
   clickMacro: Option[String])(implicit request: RequestHeader)
