+++ commercial/app/views/masterclasses/masterclassesBlended.scala.html
@@ -1,4 +1,4 @@
-@(event: model.commercial.events.Masterclass, clickMacro: Option[String])
+@(event: model.commercial.Masterclass, clickMacro: Option[String])
 
 @import conf.Configuration
