+++ commercial/app/views/soulmates/soulmateCard.scala.html
@@ -1,4 +1,4 @@
-@(soulmate: model.commercial.soulmates.Member,
+@(soulmate: model.commercial.Member,
   clickMacro: Option[String],
   optClassNames: Option[Seq[String]] = None)
