+++ commercial/app/views/soulmates/soulmates.scala.html
@@ -1,4 +1,4 @@
-@(members: Seq[model.commercial.soulmates.Member],
+@(members: Seq[model.commercial.Member],
   omnitureId: Option[String],
   clickMacro: Option[String])(implicit request: RequestHeader)
