+++ commercial/app/views/soulmates/soulmatesBlended.scala.html
@@ -1,4 +1,4 @@
-@(members: Seq[model.commercial.soulmates.Member],
+@(members: Seq[model.commercial.Member],
   clickMacro: Option[String])
 
 @import conf.Configuration
