+++ commercial/app/views/travel/travel.scala.html
@@ -1,4 +1,4 @@
-@(offers: Seq[model.commercial.travel.TravelOffer],
+@(offers: Seq[model.commercial.TravelOffer],
   omnitureId: Option[String],
   clickMacro: Option[String],
   isProminent: Boolean = false)(implicit request: RequestHeader)
