+++ commercial/app/views/travel/travelBlended.scala.html
@@ -1,4 +1,4 @@
-@(offer: model.commercial.travel.TravelOffer, clickMacro: Option[String])
+@(offer: model.commercial.TravelOffer, clickMacro: Option[String])
 
 @import conf.Configuration
