+++ commercial/app/views/travel/travelCard.scala.html
@@ -1,4 +1,4 @@
-@(offer: model.commercial.travel.TravelOffer,
+@(offer: model.commercial.TravelOffer,
   clickMacro: Option[String],
   optClassNames: Option[Seq[String]] = None)
