+++ commercial/app/views/travel/travelLargeCard.scala.html
@@ -1,4 +1,4 @@
-@(offer: model.commercial.travel.TravelOffer,
+@(offer: model.commercial.TravelOffer,
   clickMacro: Option[String],
   optAdvertClassNames: Option[Seq[String]] = None)
